Add React lazy loading and buttons hover
