+++ src/components/About.jsx
@@ -13,6 +13,7 @@ import ruby from '../assets/ruby.png';
 import rails from '../assets/rails.png';
 import tailwind from '../assets/tailwind.png';
 import bootstrap from '../assets/bootstrap.png';
+import resume from '../assets/resume.pdf';
 
 
 
@@ -45,7 +46,11 @@ const About = () => {
           <li><img src={bootstrap} alt="Bootstrap icon" /> Bootstrap</li>
         </ul>
           <div className="about-resume">
-            <button type="button" className="btn btn--green">Download my resume</button>
+            <a href={resume} download>
+              <button type="button" className="btn btn--green">
+                Download my resume
+              </button>
+            </a>
           </div>
       </div>
     </>
